Introduce a MAC Formatter
Preserve existing MAC formatting functionality from the existing PHP
backend.
We want to ensure that the MAC is in a good format before persisting it
in the session table.

There are some edge cases where we still end up with an invalid MAC.
I have documented this with RSpec examples.
Leaving this as is for now, can revisit later, but ultimately we expect
that the supplied MAC address is in a good format.