Make parameters in post-auth route optional
It is not garuanteed that we will receive all the parameters from the
radius servers.
Based on the functionality of the current backend, we will persist empty values.

This even applies to the username which doesn't make much sense because
it cannot be used to identify the indivdual that was authorised.

This functionality can be changed once this current implementation has
been proven to work in production.