Ensure either the ap or the building identifier is set
The "Called Station ID" param can contain either an access point value
or a building identifier.  Once formatted, we can see if it is 17 chars
long and know that it is an acces point identifier.  If it is not blank
and not a valid MAC, we can assume that it is a building identifier.