Fix bug Called Station ID saved as an Access Point
When we receive a Called Station ID we check whether it is a valid MAC.
If it is, save it as an Access Point, otherwise try save it as a
building identifier.

When checking whether it's a valid Access Point, we need to check its
formatted version, which this code wasn't doing previously.