Record start time for a session
This functionality was missing, so add it with this commit.
When we create a session record, we populate the start time with the
current time.

Also fix date format with strftime in for userdetails last_login, it was
using %y which is a date like 18, not 2018. MYSQL overlooked this and
saved it correctly anyway, but fix the code for consistency.