Generate statistics for sessions from the database
This is the basic functionality to get the statistics from the database
in raw form.

A later commit will introduce a presenter to transform
these results into something the performance platform will accept.

Nothing uses this yet, a later commit will introduce the rake task to
send these over to the performance platform.