Increase the read timeout for Sequel gem
We're running into issue where a long running sql query (roughly 10
minutes) is being cancelled by the ruby client.

We suspect that increasing this timeout value will prevent this from
happening but needs to be tested against the production dataset.

Set the timeout to 9999, which should be more than enough time