Keep less session data in the database
The maximum amount of session data we need to do reporting is 1 month.
Originally we kept more than this to facilitate with criminal investigations,
this can still be done using Cloudwatch instead of the database.

This should reduce the size of the sessions table by 2 thirds, making it
less difficult to manage and enabling us to switch to a less expensive
RDS image down the line.