Never return 404 (#108)
Given Logging has been removed from the "critical path", we never want
it to influence the result of the radius request based on a logging
operation.

If the logging API returns a 404, it will implicitly change an
Access-Accept to an Access-Reject.  We never want this to happen.

Always return a 204, indicating to the radius servers that everything
has gone well.