Adjust Batch deletion of old sessions to reduce RDS lag
Currently when the deletion of old sessions happens, a lag
in RDS replication occurs, which in turn fires a cloudwatch
alarm.  As it self heals, this alarm is ignored but it does
create uneccessary noise which we have all become numb to.

We beleive the cause is because of how large the query is,
and that it creates a single large transaction log which
needs to be first copied across and fully replayed on the
replica, and during this replay the lag is created.

This PR breaks the session deletion query into batches of
100, with the intent to allow each batch to be replicated
and replayed on to the replica whilst other batches are
still being created on the master, thus dampening the
effect of the replication and reducing lag.

Pair: @smford & @rjbaker

created when the RDS replica