feature: add daily reporting for the publish_statistics task
The monthly and weekly tasks are identical except for the SQL keyword
used to chunk up the data. Leverage this with a single task redefined
for each period (day/week/month) through a mapping that also holds the
corresponding adverbial (daily/weekly/monthly) for use in the task
name (manually inferring it is awkward: `"#{period}ly"` doesn't work
for `day`).