Use rubocop-govuk instead of govuk-lint
Govuk-lint is deprecated, so use rubocop-govuk instead.

Note that it has been about two years since the govuk-lint dependency
in this project was updated, so making this change will introduce a
large number of linting failures, which will be fixed in subsequent
commits in the same PR as this one.