Allow setting the DB read_timeout to context specific values
This is quite crude in my opinion, since the DB constant gets in the
way, and I'm not sure there's a way of changing the read_timeout for
an existing database connection.

Anyway, this seems to work. The timeout when running a server is
lower, since the queries need to finish before requests time out. The
timeout for the Rake tasks is longer though, since some of the queries
take ages.