Update active_users query to not use the 'Date' function
Running the query previously proved to be extremely slow. This is
because the Sessions table contains > 100 million rows and the query
used the "Date" function to select rows with a start attribute within
the specified day. This means that for each of the >100 million rows
the date function is applied which also makes it impossible for the
database to use the relevant index.

The change removes the 'Date' function from the query and instead
performs a direct comparison.

An extra test was introduced to ensure we select the correct Session
rows.