Optimise the calculation of roaming users
The sessions table is getting very large and is slowing down the computation
of roaming users.

It is optimised by omitting the time interval which is hard to write and index
for and is replaced by selecting for dates greater than a day/week/month ago.o

An index is added to the sessions database to optimise this query, which now
runs in around 6 minutes on an M1 mac laptop.