Update the smoketest cleanup code
The smoke tests have recently changed. They now take care of
removing users that have been generated during the tests.

The Session data however still has to be removed. This is now
done by siteIP and a new environment variable is introduced
'SMOKE_TEST_IPS' that lists all the site ips that the smoke tests
use so that the correct sessions will be deleted.